Deleted unused TimingSensitivityCollectionViewController
